modify make_screenshots.py to invoke WaxSim correctly for iPad
WaxSim needs to have a "-f ipad" command-line switch when running for iPad…it is not sufficient to modify the NSUserDefaults for com.apple.iphonesimulator.

I'm not much of a Python programmer, so there is probably a more elegant way to do what I'm doing.

Note that the sample project is phone-only, so the screenshots will not work correctly if add iPad to the device list.  However, you will be able to see that the simulator is correctly launched as iPad.

I have verified on universal apps that the screenshots do work correctly on iPad with this script change.